+++ crates/ruma-common/CHANGELOG.md
@@ -12,6 +12,7 @@ Improvements:
   cases where we receive a HTTP header with an unexpected value.
 - Implement `Eq`/`Hash`/`PartialEq` for `ThirdPartyIdentifier`, to check whether 
   a `ThirdPartyIdentifier` has already been added by another user.
+- Add `MatrixVersion::V1_11`
 
 # 0.13.0